[refactor/#63] search 페이지 컴포넌트 구조 개선
